Write integration tests inside golang
I think it will be easier to maintain a suite of integration tests written in golang, especially taking Windows into account where bash and GNU tools are going to be trickier.

This will allow them to run on any platform golang does, and will reduce the complexity of contributing more.

To update the golden files, `UPDATE_GOLDEN=anything go test -tags integration ./cmd/goss`. The person doing that basically then has to inspect the diff to see that they're happy. This also makes it very obvious during code-review when user-facing things change.

Note: the VERSION from help is omitted because that's injected in by the release build -ldflags.