Make goss compile on windows. (#2)
I've extracted the unix-only part to file_unix.go and windows part to file_windows.go. I haven't found an easy way to support owners/groups on windows, so I've left it empty.

TODO: figure out how to write and run integration tests on windows, but probably not this pr.