Validate checksums in installer
There exist checksums for both `goss` and `dgoss` since version 0.3.11,
so those should also be used to validate the downloads.
This only works since 0.3.13 though, as the filename in the checksum
file changed.

Also use the same version for dgoss than for goss, instead of always
latest master for dgoss.