chore: first pass at dependency version updates (#739)
* chore: bump Masterminds/goutils to 1.1.1

* chore: bump fatih/color and its dependent modules

This bumps the release version of the color module, which itself
includes a number of downstream improvements. None of this appears
to include any breaking changes for us.

* chore: bump davecgh/go-spew to v1.1.1

This does not appear to have any breaking changes for us, should
be a simple update.

* chore: bump fsnotify to v1.5.1

This has some notable improvements to its implementation but none
that appear to require any changes from us. There was a potential
breaking change around symlinks in the 1.5.0 release but it was
reverted for 1.5.1.

* chore: bump version of xstrings to 1.3.2

Minor fixes that don't introduce any impact for goss.

* chore: bump miekg/dns to 1.1.45

This is a tougher one to read, as there's no real changelog to
speak of. However, near as I can tell, the commit history
between our current version and 1.1.45 doesn't appear to include
anything that would represent a breaking change or nasty hidden
surprises to any of our expectations.

* chore: bump reflections module to 1.0.1

This looks like a big jump, but is just updating to more current go
standards with no API changes.

* chore: bump objx module to 0.3.0

This largely reflects updates to go versions and migration to
module use. There are some additional improvements but nothing
appears to be a breaking change for goss.

* chore: bump testify to v1.7.0

The summary on this is vague, but looking at the commits it appears
to be a safe update.

* chore: bump goldmark to 1.4.4

All fixes and minor improvements, including bumping the commonmark
version that's supported.

* chore: bump x/mod support to 0.5.1

* chore: bump x/text to v0.3.7

* chore: bump x/tools to v0.1.9

* chore: run go mod tidy

* chore: bump google/uuid to 1.3.0

* chore: bump copystructure and reflectwalk versions

* chore: bump tail version

* chore: bump golang/x modules

* chore: bump x/crypto

* chore: bump check.v1 and yaml.v3 modules