Add support for `podman` as the container backend
The `CONTAINER_RUNTIME` environment variable can be used to control container
runtime. Can be `docker` or `podman`. Defaults to `docker`.

To be able to use `podman` as the container backend `podman run` needs to run
a command which does not finish. Commonly `sleep infinity` is used for that.

Provide enough documentation to understand the feature for the users.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>