improve handling of errors in encoded TestResults
Go error does not marshal to a string and cannot be unmarshalled
but the TestResult has a error in it and is being marshalled by
StructuredResult.

Here we create a basic string based error implementation that
goes to json and back cleanly but we make sure tests for err==nil
continues to work.

Signed-off-by: R.I.Pienaar <rip@devco.net>