Changed: Goss add command default to empty string
Changed the behavior of `goss add` on a command, if the value of stdout or stderr is
empty, it will be defaulted to an empty string instead of an empty
array. The behavior remains unchanged when there is output; the value
will be an array.

This approach provides the following benefits:
* Tests will fail if an empty string receives a value.
* Arrays allows for simpler deletion of lines that are not relevant
to the test assertion.

closes #820