add more labels to prometheus output
This adds a 'resource_id' label to each test-specific metric if the 'verbose' format option was enabled.

Since we only have access to format options in the 'Output' method, we are creating metrics during the first run of that method.

Metrics can only be registered once for any given set of labels, so we create a custom registry and use the existence of that as a check whether we need to create/register metrics. This has the added benefit of simplifying the test a tiny bit, and it exposes even less metrics on the /health endpoint.

fixes #862

Signed-off-by: Sebastian Hoß <seb@xn--ho-hia.de>