Make cli argument `vars` to be slice of string
This change allows passing multiple variables files as input to
CLI. Order of files is important as overlapped non-empty keys in yaml or json
files will be overwritten by subsequent file.

This change is backwards compatible as CLI library makes the slice
definition to be repetition of the --vars argument. Using goss with
single file produces same outputs with this code change as it was before.

Usage:
```
./goss --vars file1.yaml --vars file2.yaml validate
```

Issue ref: https://github.com/goss-org/goss/issues/1023

Signed-off-by: Ondrej Vasko <ondrej.vaskoo@gmail.com>