Update dependencies
Some of the messages the integration tests expect change in more recent
versions of Go.

Further, the .editorconfig file leads to trailing whitespace being
removed. It's a bit of a coin-toss as to whether the right solution here
is to override whitespace trimming for just the `testcase` directory or
to instead perform some normalization of the expected and actual strings
to trim this whitespace, so I did both.